랭킹(rating) 올림 처리로 수정한다. (#115)
fix: ceil rating

+++ domain/src/main/kotlin/com/gotchai/domain/user/dto/result/GetUserRankingResult.kt
@@ -4,12 +4,12 @@ import com.gotchai.domain.user.entity.Profile
 
 data class GetUserRankingResult(
     val name: String,
-    val rating: Double
+    val rating: Int
 ) {
     companion object {
         fun of(
             profile: Profile,
-            rating: Double
+            rating: Int
         ) = GetUserRankingResult(
             name = profile.nickname,
             rating = rating
